Add support for device query
The id is now queried as a UUID, and the Astarte device id is exposed as a
calculation.

This also include a massive rework of the data layer internals to make them much
more general and add some more comments. They should now (hopefully) work
without modifications also with other resources.